Move some top-level items to `apollo_compiler::validation`
