+++ crates/apollo-compiler/src/executable/validation.rs
@@ -3,8 +3,8 @@ use super::FieldSet;
 use crate::ast;
 use crate::validation::Details;
 use crate::validation::DiagnosticList;
+use crate::validation::FileId;
 use crate::ExecutableDocument;
-use crate::FileId;
 use crate::InputDatabase;
 use crate::Schema;
 use crate::ValidationDatabase;
