+++ crates/apollo-compiler/src/node_str.rs
@@ -1,6 +1,6 @@
 #![allow(unstable_name_collisions)] // for `sptr::Strict`
 
-use crate::NodeLocation;
+use crate::validation::NodeLocation;
 use sptr::Strict;
 use std::marker::PhantomData;
 use std::mem::size_of;
