+++ crates/apollo-compiler/src/parser.rs
@@ -5,9 +5,9 @@ use crate::executable;
 use crate::schema::SchemaBuilder;
 use crate::validation::Details;
 use crate::validation::DiagnosticList;
+use crate::validation::FileId;
+use crate::validation::NodeLocation;
 use crate::ExecutableDocument;
-use crate::FileId;
-use crate::NodeLocation;
 use crate::Schema;
 use indexmap::IndexMap;
 use std::path::Path;
